Add repository packages
