+++ language-server/modules/langserver-core/src/test/resources/completion/import_decl/source/repository_projects/local_project1/main.bal
