+++ language-server/modules/langserver-core/src/test/resources/completion/import_decl/source/repository_projects/local_project2/Ballerina.toml
@@ -1,0 +1,4 @@
+[package]
+org="test"
+name="local_project2"
+version="0.0.1"
