+++ language-server/modules/langserver-core/src/test/resources/completion/import_decl/source/repository_projects/project2/Ballerina.toml
@@ -1,0 +1,4 @@
+[package]
+org="test"
+name="project2"
+version="0.0.1"
