+++ language-server/modules/langserver-core/src/test/resources/completion/import_decl/source/repository_projects/project2/main.bal
