--- language-server/modules/langserver-core/src/test/resources/completion/import_decl_mock/source/projects/local_project2/.gitignore
@@ -1,1 +1,0 @@
-target
