--- language-server/modules/langserver-core/src/test/resources/completion/import_decl_mock/source/projects/project2/.devcontainer.json
@@ -1,4 +1,0 @@
-{
-    "image": "ballerina/ballerina-devcontainer:2201.0.1-rc3.1",
-    "extensions": ["WSO2.ballerina"],
-}
