--- language-server/modules/langserver-core/src/test/resources/completion/import_decl_mock/source/projects/project2/Ballerina.toml
@@ -1,5 +1,0 @@
-[package]
-distribution = "2201.0.1-rc3.1"
-
-[build-options]
-observabilityIncluded = true
